update middleware structure
